requirements.txt
ansicon==1.89.0
arrow==1.0.3
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
blessed==1.18.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
croniter==1.0.8
cryptography==3.4.6
daphne==3.0.1
Django==3.1.9
django-bootstrap-datepicker-plus==3.0.5
django-picklefield==3.0.1
django-q==1.3.5
django-widget-tweaks==1.4.8
future==0.18.2
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
jinxed==1.0.1
mysqlclient==2.0.3
natsort==7.1.1
numpy==1.20.3
pandas==1.2.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
wcwidth==0.2.5
zope.interface==5.2.0
